Correctly handle more than 32767 physical sectors in one Rwabs call
For file systems with differing logical and physical sector size, a single
Rwabs call can lead to more than 32767 physical sectors to be read or
written. With this commit, the buffer address is updated correctly when
such a call is therefore split up into multiple low-level IOs.

In particular, this fixes loading very large executables with at least
16 MiB combined text and data size.

Thank you to Eero Tamminen for reporting this issue and to Roger Burrows
for finding the root cause.

+++ bios/blkdev.c
@@ -520,7 +520,7 @@ static LONG blkdev_rwabs(WORD rw, UBYTE *buf, WORD cnt, WORD recnr, WORD dev, LO
         } while(retval == CRITIC_RETRY_REQUEST);
         if (retval < 0)     /* error, retries exhausted */
             break;
-        buf += scount << psshift;
+        buf += (ULONG)scount << psshift;
         lrecnr += scount;
         lcount -= scount;
     } while(lcount > 0);
